Extracted Association Tests to own class
